requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==3.2.14
django-form-surveys==1.6.2
django-nested-admin==3.4.0
django-resized==1.0.1
easy-thumbnails==2.7
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.3
python-monkey-business==1.0.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
whitenoise==6.2.0
